+++ app/src/main/java/com/decagonhq/decapay/feature/edit_profile/data/network/model/EditProfileApi.kt
@@ -1,8 +1,4 @@
-package com.decagonhq.decapay.feature.edit_profile.data.network.api
-
-import com.decagonhq.decapay.feature.edit_profile.data.network.model.EditProfileRequestBody
-import com.decagonhq.decapay.feature.edit_profile.data.network.model.EditProfileResponse
-import com.decagonhq.decapay.feature.edit_profile.data.network.model.UserProfileResponse
+package com.decagonhq.decapay.feature.edit_profile.data.network.model
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.PUT
