+++ app/src/main/java/com/decagonhq/decapay/feature/edit_profile/data/repository/EditProfileRepositoryImpl.kt
@@ -3,11 +3,16 @@ package com.decagonhq.decapay.feature.edit_profile.data.repository
 import com.decagonhq.decapay.feature.edit_profile.data.network.api.EditProfileApi
 import com.decagonhq.decapay.feature.edit_profile.data.network.model.EditProfileRequestBody
 import com.decagonhq.decapay.feature.edit_profile.data.network.model.EditProfileResponse
+import com.decagonhq.decapay.feature.edit_profile.data.network.model.UserProfileResponse
 import com.decagonhq.decapay.feature.edit_profile.domain.repository.EditProfileRepository
 import javax.inject.Inject
 
 class EditProfileRepositoryImpl @Inject constructor(private  val editProfileApi: EditProfileApi): EditProfileRepository {
     override suspend fun editProfile(editProfileRequestBody: EditProfileRequestBody): EditProfileResponse {
         return  editProfileApi.editProfile(editProfileRequestBody)
+    }
+
+    override suspend fun getUseProfile(): UserProfileResponse {
+        return editProfileApi.getUserProfile()
     }
 }
