+++ app/src/main/java/com/decagonhq/decapay/feature/edit_profile/domain/usecase/UserProfileUsecase.kt
@@ -1,11 +1,9 @@
-package com.decagonhq.decapay.feature.userprofile.domain.usecase
+package com.decagonhq.decapay.feature.edit_profile.domain.usecase
 
 import com.decagonhq.decapay.common.utils.errorhelper.ExceptionHandler
 import com.decagonhq.decapay.common.utils.resource.Resource
-import com.decagonhq.decapay.feature.listbudgetcategories.data.network.model.BudgetCategoriesResponse
-import com.decagonhq.decapay.feature.listbudgetcategories.domain.repository.BudgetCategoryListRepository
-import com.decagonhq.decapay.feature.userprofile.data.network.model.UserProfileResponse
-import com.decagonhq.decapay.feature.userprofile.domain.repository.UserProfileRepository
+import com.decagonhq.decapay.feature.edit_profile.data.network.model.UserProfileResponse
+import com.decagonhq.decapay.feature.edit_profile.domain.repository.EditProfileRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import retrofit2.HttpException
@@ -13,14 +11,14 @@ import java.io.IOException
 import javax.inject.Inject
 
 class UserProfileUsecase @Inject constructor(
-    private val userProfileRepository: UserProfileRepository,
+    private val editProfileRepository: EditProfileRepository,
     private val exceptionHandler: ExceptionHandler
 ) {
     operator fun invoke(): Flow<Resource<UserProfileResponse>> = flow {
 
         try {
             emit(Resource.Loading())
-            val response = userProfileRepository.getUseProfile()
+            val response = editProfileRepository.getUseProfile()
             emit(Resource.Success(response))
         } catch (e: HttpException) {
             val message = exceptionHandler.parse(e)
