--- app/src/main/java/com/decagonhq/decapay/feature/userprofile/data/network/api/UserProfileApi.kt
@@ -1,12 +1,0 @@
-package com.decagonhq.decapay.feature.userprofile.data.network.api
-
-import com.decagonhq.decapay.feature.userprofile.data.network.model.UserProfileResponse
-import retrofit2.http.GET
-
-interface UserProfileApi {
-
-    @GET("api/v1/profile")
-    suspend fun getUserProfile(
-
-    ): UserProfileResponse
-}
