--- app/src/main/java/com/decagonhq/decapay/feature/userprofile/domain/repository/UserProfileRepository.kt
@@ -1,10 +1,0 @@
-package com.decagonhq.decapay.feature.userprofile.domain.repository
-
-import com.decagonhq.decapay.feature.userprofile.data.network.model.UserProfileResponse
-
-interface UserProfileRepository {
-
-    suspend fun  getUseProfile(): UserProfileResponse
-
-
-}
