+++ app/src/main/res/layout/fragment_edit_profile.xml
@@ -40,7 +40,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:fontFamily="@font/roboto_medium"
-                android:text="@string/edit_profile"
+                android:text="@string/my_profile"
                 android:textSize="@dimen/create_budget_fragment_toolbar_title_txsiz"
                 android:textStyle="bold"
                 app:layout_constraintEnd_toEndOf="parent"
@@ -188,7 +188,7 @@ android:id="@+id/edit_profile_last_name_tv"
         app:backgroundTint="@color/decapay_green"
         android:layout_marginBottom="30dp"
         android:foreground="?attr/selectableItemBackgroundBorderless"
-        android:text="@string/submit"
+        android:text="@string/edit_profile"
         android:textAllCaps="false"
         app:layout_constraintHorizontal_bias="0.6"
         />
