+++ app/src/main/res/navigation/authentication_nav_graph.xml
@@ -121,11 +121,6 @@
         android:name="com.decagonhq.decapay.feature.usersettings.presentation.UserSettingsFragment"
         android:label="UserSettingsFragment" />
     <fragment
-        android:id="@+id/userProfileFragment"
-        android:name="com.decagonhq.decapay.feature.userprofile.presentation.UserProfileFragment"
-        android:label="fragment_user_profile"
-        tools:layout="@layout/fragment_user_profile" />
-    <fragment
         android:id="@+id/editProfileFragment"
         android:name="com.decagonhq.decapay.feature.edit_profile.presentation.EditProfileFragment"
         android:label="fragment_edit_profile"
